+++ skbio/diversity/alpha/__init__.py
@@ -13,7 +13,9 @@ where each count is, for example, the number of observations of a particular
 Operational Taxonomic Unit, or OTU. We use the term "OTU" here very loosely, as
 these could be counts of any type of feature/observation (e.g., bacterial
 species). We'll refer to this vector as the *counts vector* or simply *counts*
-throughout the documentation.
+throughout the documentation. Some of these metrics incorporate phylogeny.
+These metrics take a tree (``skbio.TreeNode``) and a list of OTU ids mapping
+the counts to tips in the tree, in addition to the counts vector.
 
 The counts vector must be one-dimensional and contain integers representing the
 number of individuals seen (or *counted*) for a particular OTU. Negative values
@@ -95,6 +97,7 @@ Functions
 Examples
 --------
 
+>>> from skbio.diversity.alpha import observed_otus
 >>> import numpy as np
 
 Assume we have the following abundance data for a sample, represented as a
@@ -124,10 +127,28 @@ exactly one time (i.e., are *singleton* OTUs), and ``doubles``, which tells us
 how many OTUs are observed exactly two times (i.e., are *doubleton* OTUs).
 Let's see how many singletons and doubletons there are in the sample:
 
+>>> from skbio.diversity.alpha import singles, doubles
 >>> singles(counts)
 2
 >>> doubles(counts)
 1
+
+Phylogenetic diversity metrics additionally incorporate the relative
+relatedness of the OTUs in the calculation, and therefore require a tree and
+a mapping of counts to OTU (tip) ids in the tree. Here we'll apply Faith's
+Phylogenetic Diversity (PD) metric to the sample:
+
+>>> from skbio import TreeNode
+>>> from skbio.diversity.alpha import faith_pd
+>>> from io import StringIO
+>>> tree = TreeNode.read(StringIO(
+...                      u'(((((OTU1:0.5,OTU2:0.5):0.5,OTU3:1.0):1.0):0.0,'
+...                      u'(OTU4:0.75,(OTU5:0.5,((OTU6:0.33,OTU7:0.62):0.5,'
+...                      u'OTU8:0.5):0.5):0.5):1.25):0.0)root;'))
+>>> otu_ids = ['OTU1', 'OTU2', 'OTU3', 'OTU4', 'OTU5', 'OTU6', 'OTU7', 'OTU8']
+>>> faith_pd(counts, otu_ids, tree)
+6.95
+
 
 """